+++ smeem-api/src/main/java/com/smeem/api/diary/dto/request/DiaryListGetServiceRequest.java
@@ -1,4 +1,4 @@
-package com.smeem.api.diary.service.dto.request;
+package com.smeem.api.diary.dto.request;
 
 import lombok.Builder;
