+++ smeem-api/src/main/java/com/smeem/api/diary/dto/request/DiaryModifyServiceRequest.java
@@ -1,6 +1,5 @@
-package com.smeem.api.diary.service.dto.request;
+package com.smeem.api.diary.dto.request;
 
-import com.smeem.api.diary.controller.dto.request.DiaryModifyRequest;
 import lombok.Builder;
 
 import static lombok.AccessLevel.PRIVATE;
