+++ smeem-api/src/main/java/com/smeem/api/diary/dto/response/DiaryCreateServiceResponse.java
@@ -1,4 +1,4 @@
-package com.smeem.api.diary.service.dto.response;
+package com.smeem.api.diary.dto.response;
 
 import com.smeem.api.badge.controller.dto.response.AcquiredBadgeResponseDTO;
 import com.smeem.domain.badge.model.Badge;
