+++ smeem-api/src/test/java/com/smeem/api/controller/DiaryControllerTest.java
@@ -5,12 +5,12 @@ import com.smeem.api.badge.controller.dto.response.AcquiredBadgeResponseDTO;
 import com.smeem.api.common.ApiResponseUtil;
 import com.smeem.api.common.BaseResponse;
 import com.smeem.api.diary.controller.DiaryController;
-import com.smeem.api.diary.controller.dto.request.DiaryCreateRequest;
-import com.smeem.api.diary.controller.dto.request.DiaryModifyRequest;
-import com.smeem.api.diary.service.dto.response.DiaryCreateServiceResponse;
-import com.smeem.api.diary.service.dto.response.DiaryListGetServiceResponse;
-import com.smeem.api.diary.service.dto.response.DiaryGetServiceResponse;
-import com.smeem.api.diary.service.dto.response.DiaryListGetServiceResponse.DiaryResponse;
+import com.smeem.api.diary.dto.request.DiaryCreateRequest;
+import com.smeem.api.diary.dto.request.DiaryModifyRequest;
+import com.smeem.api.diary.dto.response.DiaryCreateServiceResponse;
+import com.smeem.api.diary.dto.response.DiaryListGetServiceResponse;
+import com.smeem.api.diary.dto.response.DiaryGetServiceResponse;
+import com.smeem.api.diary.dto.response.DiaryListGetServiceResponse.DiaryResponse;
 import lombok.val;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.ResponseEntity;
